fix multiple issues related to character ratings
