+++ tests/user.test.ts
@@ -1827,404 +1827,6 @@ Deno.test('/image', async (test) => {
   });
 });
 
-Deno.test('media characters', async (test) => {
-  await test.step('normal', async () => {
-    const characterStub = stub(
-      packs,
-      'mediaCharacters',
-      () =>
-        Promise.resolve({
-          next: true,
-          media: {
-            id: '1',
-            packId: 'pack-id',
-            type: MediaType.Anime,
-            title: {
-              english: 'title',
-            },
-          },
-          character: {
-            id: '2',
-            packId: 'pack-id',
-            name: {
-              english: 'name',
-            },
-          },
-        }),
-    );
-
-    const fetchStub = stub(
-      globalThis,
-      'fetch',
-      () => ({
-        ok: true,
-        text: (() =>
-          Promise.resolve(JSON.stringify({
-            data: {},
-          }))),
-      } as any),
-    );
-
-    const listStub = stub(
-      packs,
-      'all',
-      () => Promise.resolve([]),
-    );
-
-    const isDisabledStub = stub(packs, 'isDisabled', () => false);
-
-    try {
-      const message = await search.mediaCharacters({
-        id: 'pack-id:1',
-        guildId: 'guild_id',
-        channelId: 'channel_id',
-        index: 0,
-      });
-
-      assertEquals(message.json(), {
-        type: 4,
-        data: {
-          attachments: [],
-          components: [{
-            type: 1,
-            components: [
-              {
-                custom_id: '_',
-                disabled: true,
-                label: '1',
-                style: 2,
-                type: 2,
-              },
-              {
-                custom_id: 'mcharacters=pack-id:1=1=next',
-                label: 'Next',
-                style: 2,
-                type: 2,
-              },
-              {
-                custom_id: 'like=pack-id:2',
-                label: '/like',
-                style: 2,
-                type: 2,
-              },
-              {
-                custom_id: 'media=pack-id:1',
-                label: '/anime',
-                style: 2,
-                type: 2,
-              },
-            ],
-          }],
-          embeds: [
-            {
-              type: 'rich',
-              description:
-                '<:star:1061016362832642098><:no_star:1061016360190222466><:no_star:1061016360190222466><:no_star:1061016360190222466><:no_star:1061016360190222466>',
-              fields: [
-                {
-                  name: 'name',
-                  value: '\u200B',
-                },
-              ],
-              image: {
-                url: 'undefined/external/',
-              },
-            },
-          ],
-        },
-      });
-    } finally {
-      characterStub.restore();
-      fetchStub.restore();
-      listStub.restore();
-      isDisabledStub.restore();
-    }
-  });
-
-  await test.step('with owner', async () => {
-    const characterStub = stub(
-      packs,
-      'mediaCharacters',
-      () =>
-        Promise.resolve({
-          next: true,
-          media: {
-            id: '1',
-            packId: 'pack-id',
-            type: MediaType.Manga,
-            title: {
-              english: 'title',
-            },
-          },
-          character: {
-            id: '2',
-            packId: 'pack-id',
-            name: {
-              english: 'name',
-            },
-          },
-        }),
-    );
-
-    const fetchStub = stub(
-      globalThis,
-      'fetch',
-      () => ({
-        ok: true,
-        text: (() =>
-          Promise.resolve(JSON.stringify({
-            data: {
-              findCharacter: {
-                user: {
-                  id: 'user_id',
-                },
-                id: 'character_id',
-                mediaId: 'media_id',
-                rating: 3,
-              },
-            },
-          }))),
-      } as any),
-    );
-
-    const listStub = stub(
-      packs,
-      'all',
-      () => Promise.resolve([]),
-    );
-
-    const isDisabledStub = stub(packs, 'isDisabled', () => false);
-
-    try {
-      const message = await search.mediaCharacters({
-        id: 'pack-id:1',
-        guildId: 'guild_id',
-        channelId: 'channel_id',
-        index: 0,
-      });
-
-      assertEquals(message.json(), {
-        type: 4,
-        data: {
-          attachments: [],
-          components: [{
-            type: 1,
-            components: [
-              {
-                custom_id: '_',
-                disabled: true,
-                label: '1',
-                style: 2,
-                type: 2,
-              },
-              {
-                custom_id: 'mcharacters=pack-id:1=1=next',
-                label: 'Next',
-                style: 2,
-                type: 2,
-              },
-              {
-                custom_id: 'like=pack-id:2',
-                label: '/like',
-                style: 2,
-                type: 2,
-              },
-              {
-                custom_id: 'media=pack-id:1',
-                label: '/manga',
-                style: 2,
-                type: 2,
-              },
-            ],
-          }],
-          embeds: [
-            {
-              type: 'rich',
-              description:
-                '<@user_id>\n\n<:star:1061016362832642098><:star:1061016362832642098><:star:1061016362832642098><:no_star:1061016360190222466><:no_star:1061016360190222466>',
-              fields: [
-                {
-                  name: 'name',
-                  value: '\u200B',
-                },
-              ],
-              image: {
-                url: 'undefined/external/',
-              },
-            },
-          ],
-        },
-      });
-
-      assertSpyCalls(fetchStub, 1);
-    } finally {
-      characterStub.restore();
-      fetchStub.restore();
-      listStub.restore();
-      isDisabledStub.restore();
-    }
-  });
-
-  await test.step('disabled character', async () => {
-    const characterStub = stub(
-      packs,
-      'mediaCharacters',
-      () =>
-        Promise.resolve({
-          next: false,
-          media: {
-            id: '1',
-            type: MediaType.Anime,
-            title: {
-              english: 'title',
-            },
-          },
-          character: {
-            id: '2',
-            name: {
-              english: 'name',
-            },
-          },
-        }),
-    );
-
-    const listStub = stub(
-      packs,
-      'all',
-      () => Promise.resolve([]),
-    );
-
-    const isDisabledStub = stub(packs, 'isDisabled', () => true);
-
-    try {
-      await assertRejects(
-        async () =>
-          await search.mediaCharacters({
-            id: 'pack-id:1',
-            guildId: 'guild_id',
-            channelId: 'channel_id',
-            index: 0,
-          }),
-        NonFetalError,
-        'This character was removed or disabled',
-      );
-    } finally {
-      characterStub.restore();
-      listStub.restore();
-      isDisabledStub.restore();
-    }
-  });
-
-  await test.step('no characters', async () => {
-    const characterStub = stub(
-      packs,
-      'mediaCharacters',
-      () =>
-        Promise.resolve({
-          next: false,
-          media: {
-            id: '1',
-            type: MediaType.Anime,
-            title: {
-              english: 'title',
-            },
-          },
-        }),
-    );
-
-    const listStub = stub(
-      packs,
-      'all',
-      () => Promise.resolve([]),
-    );
-
-    try {
-      await assertRejects(
-        async () =>
-          await search.mediaCharacters({
-            id: 'pack-id:1',
-            guildId: 'guild_id',
-            channelId: 'channel_id',
-            index: 0,
-          }),
-        NonFetalError,
-        'title contains no characters',
-      );
-    } finally {
-      characterStub.restore();
-      listStub.restore();
-    }
-  });
-
-  await test.step('no more characters', async () => {
-    const characterStub = stub(
-      packs,
-      'mediaCharacters',
-      () =>
-        Promise.resolve({
-          next: false,
-          media: {
-            id: '1',
-            type: MediaType.Anime,
-            title: {
-              english: 'title',
-            },
-          },
-        }),
-    );
-
-    const listStub = stub(
-      packs,
-      'all',
-      () => Promise.resolve([]),
-    );
-
-    try {
-      await assertRejects(
-        async () =>
-          await search.mediaCharacters({
-            id: 'pack-id:1',
-            guildId: 'guild_id',
-            channelId: 'channel_id',
-            index: 1,
-          }),
-        NonFetalError,
-        'title contains no more characters',
-      );
-    } finally {
-      characterStub.restore();
-      listStub.restore();
-    }
-  });
-
-  await test.step('not found', async () => {
-    const characterStub = stub(
-      packs,
-      'mediaCharacters',
-      () =>
-        Promise.resolve({
-          next: false,
-        }),
-    );
-
-    try {
-      await assertRejects(
-        async () =>
-          await search.mediaCharacters({
-            id: 'pack-id:1',
-            guildId: 'guild_id',
-            channelId: 'channel_id',
-            index: 0,
-          }),
-        Error,
-        '404',
-      );
-    } finally {
-      characterStub.restore();
-    }
-  });
-});
-
 Deno.test('/collection stars', async (test) => {
   await test.step('normal', async () => {
     const media: AniListMedia[] = [
